--- requirements.txt
@@ -1,52 +1,0 @@
-certifi==2020.4.5.1
-chardet==3.0.4
-clean-text==0.1.1
-cleantext==1.1.3
-click==7.1.1
-cycler==0.10.0
-dash==1.4.1
-dash-core-components==1.3.1
-dash-html-components==1.0.1
-dash-renderer==1.1.2
-dash-table==4.4.1
-decorator==4.4.2
-Flask==1.1.2
-Flask-Compress==1.4.0
-ftfy==5.7
-future==0.18.2
-idna==2.9
-itsdangerous==1.1.0
-Jinja2==2.11.1
-kiwisolver==1.1.0
-MarkupSafe==1.1.1
-matplotlib==3.1.3
-mkl-fft==1.0.15
-mkl-random==1.1.0
-mkl-service==2.3.0
-networkx==2.4
-nltk==3.4.4
-numpy==1.18.1
-oauthlib==3.1.0
-pandas==1.0.3
-plotly==4.6.0
-psutil==5.7.0
-pydot==1.4.1
-pyparsing==2.4.6
-PySocks==1.7.1
-python-dateutil==2.8.1
-pytz==2019.3
-PyYAML==5.3.1
-requests==2.23.0
-requests-oauthlib==1.3.0
-retrying==1.3.3
-scipy==1.4.1
-seaborn==0.10.1
-six==1.14.0
-textblob==0.15.3
-tornado==6.0.4
-tweepy==3.8.0
-Unidecode==1.1.1
-urllib3==1.25.9
-wcwidth==0.1.9
-Werkzeug==1.0.1
-wincertstore==0.2
